wakeup gesture fix

+++ include/linux/clearpad-rhine.h
@@ -66,7 +66,7 @@ struct synaptics_button_data {
 struct synaptics_easy_wakeup_config {
 	bool gesture_enable;
 	bool large_panel;
-	u32 timeout_delay;
+	unsigned long timeout_delay;
 };
 
 struct clearpad_platform_data {
